# docs/swagger.yaml
basePath: /
definitions:
...
  blogs.BlogWithMeta:
    properties:
      blog:
        $ref: '#/definitions/models.Blog'
      commentsCount:
        type: integer
      liked:
        type: boolean
      likesCount:
        type: integer
      reposted:
        type: boolean
      repostsCount:
        type: integer
    type: object
  blogs.CreateBlogDto:
    properties:
      RepostedFromBlogId:
        example: some-other-blog-id
        type: string
      externalLink:
        example: https://example.com/related-article
        type: string
      externalLinkTitle:
        example: Related Article
        type: string
      images:
        example:
        - https://example.com/image1.jpg
        - https://example.com/image2.jpg
        items:
          type: string
        type: array
      slug:
        example: my-first-blog-post
        type: string
      text:
        example: This is the content of my first blog post.
        type: string
      title:
        example: My First Blog Post
        type: string
      url:
        example: /blogs/my-first-blog-post
        type: string
      video:
        example: https://example.com/video.mp4
        type: string
    required:
    - slug
    - text
    - title
    - url
    type: object
  blogs.CreateCommentDto:
    properties:
      image:
        example: https://example.com/comment-image.jpg
        type: string
      sticker:
        example: smiley_face
        type: string
      text:
        example: This is a great post!
        type: string
      video:
        example: https://example.com/comment-video.mp4
        type: string
    required:
    - text
    type: object
...
  blogs.LikeResponse:
    properties:
      liked:
        type: boolean
      likesCount:
        type: integer
    type: object
  blogs.MutationResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  blogs.UpdateBlogDto:
    properties:
      externalLink:
        example: https://example.com/another-related-article
        type: string
      externalLinkTitle:
        example: Another Related Article
        type: string
      images:
        example:
        - https://example.com/image3.jpg
        items:
          type: string
        type: array
      slug:
        example: my-updated-blog-post
        type: string
      text:
        example: This is the updated content of my blog post.
        type: string
      title:
        example: My Updated Blog Post
        type: string
      url:
        example: /blogs/my-updated-blog-post
        type: string
      video:
        example: https://example.com/updated-video.mp4
        type: string
    type: object
...
  models.ErrorResponse:
    description: Error response structure with a message.
    properties:
      message:
        type: string
    type: object
...
  models.PaginatedResponse:
    properties:
      data: {}
      metadata:
        $ref: '#/definitions/models.PaginationMetadata'
    type: object
...
host: localhost:5005
info:
  contact: {}
  description: This is the API documentation for the Phinex Blog Application.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Phinex Blog API
  version: "1.0"
paths:
...
  /blogs:
    get:
      consumes:
      - application/json
      description: Get all blogs with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all blogs
      tags:
      - Blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post
      parameters:
      - description: Blog object to be created
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/blogs.CreateBlogDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/blogs.MutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new blog
      tags:
      - Blogs
  /blogs/{blogId}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a blog
      tags:
      - Blogs
    get:
      consumes:
      - application/json
      description: Get a single blog by its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blogs.BlogWithMeta'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a single blog by ID
      tags:
      - Blogs
    put:
      consumes:
      - application/json
      description: Update a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      - description: Blog object to be updated
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/blogs.UpdateBlogDto'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/blogs.MutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing blog
      tags:
      - Blogs
  /blogs/{blogId}/comments:
    get:
      consumes:
      - application/json
      description: Get all comments for a specific blog with pagination
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comments for a blog
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Add a comment to a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      - description: Comment object to be created
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/blogs.CreateCommentDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/blogs.MutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a comment to a blog
      tags:
      - Comments
  /blogs/{blogId}/follows/likes:
    get:
      consumes:
      - application/json
      description: Get all likes and followed or followers that like a specific blog
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get likes and followers for a blog
      tags:
      - Blogs
  /blogs/{blogId}/likes:
    put:
      consumes:
      - application/json
      description: Like and/or unlike a blog post by its ID
      parameters:
      - description: Blog ID
        in: path
        name: blogId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blogs.LikeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like or unlike a blog
      tags:
      - Blogs
  /comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      - description: Comment object to be updated
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/blogs.CreateCommentDto'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/blogs.MutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a comment
      tags:
      - Comments
  /comments/{commentId}/likes:
    get:
      consumes:
      - application/json
      description: Get all likes for a specific comment
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Like'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get comment likes
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Like and/or unlike a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blogs.LikeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like or unlike a comment
      tags:
      - Comments
  /comments/{commentId}/replies:
    get:
      consumes:
      - application/json
      description: Get all replies for a specific comment with pagination
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get replies for a comment
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Add a reply to a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      - description: Reply object to be created
        in: body
        name: reply
        required: true
        schema:
          $ref: '#/definitions/blogs.CreateReplyDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/blogs.MutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a reply to a comment
      tags:
      - Comments
...
  /users/{userId}/blogs:
    get:
      consumes:
      - application/json
      description: Get all blogs by a specific user with pagination
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user blogs
      tags:
      - Blogs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"